backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
asyncpg
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]
python-dotenv
gunicorn
uvicorn[standard]
pandas
openpyxl
